refactor: FavoriteService 분리
