+++ src/main/java/wooteco/subway/service/member/MemberService.java
@@ -1,7 +1,6 @@
 package wooteco.subway.service.member;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.member.Favorite;
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.domain.member.MemberRepository;
 import wooteco.subway.exception.NoMemberExistException;
@@ -50,21 +49,5 @@ public class MemberService {
     public boolean loginWithForm(String email, String password) {
         Member member = findMemberByEmail(email);
         return member.checkPassword(password);
-    }
-
-    public void addFavorite(long memberId, long sourceId, long targetId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(NoMemberExistException::new);
-
-        member.addFavorite(Favorite.of(sourceId, targetId));
-        memberRepository.save(member);
-    }
-
-    public void removeFavorite(long memberId, long sourceId, long targetId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(NoMemberExistException::new);
-
-        member.removeFavorite(Favorite.of(sourceId, targetId));
-        memberRepository.save(member);
     }
 }
