+++ src/test/java/wooteco/subway/service/member/MemberServiceTest.java
@@ -1,7 +1,6 @@
 package wooteco.subway.service.member;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -13,10 +12,10 @@ import wooteco.subway.service.member.dto.LoginRequest;
 
 import java.util.Optional;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 public class MemberServiceTest {
@@ -54,32 +53,5 @@ public class MemberServiceTest {
         memberService.createToken(loginRequest);
 
         verify(jwtTokenProvider).createToken(anyString());
-    }
-
-    @DisplayName("즐겨찾기 추가 서비스")
-    @Test
-    void addFavorite() {
-        Member member = new Member(1L, TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-        when(memberRepository.findById(any())).thenReturn(Optional.of(member));
-
-        memberService.addFavorite(1L, 1L, 2L);
-
-        assertThat(member.getFavorites().getFavorites().size()).isEqualTo(1);
-
-        verify(memberRepository).save(any());
-    }
-
-    @DisplayName("즐겨찾기 제거 서비스")
-    @Test
-    void removeFavorite() {
-        Member member = new Member(1L, TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-        when(memberRepository.findById(any())).thenReturn(Optional.of(member));
-
-        memberService.addFavorite(1L, 1L, 2L);
-        memberService.removeFavorite(1L, 1L, 2L);
-
-        assertThat(member.getFavorites().getFavorites().size()).isEqualTo(0);
-
-        verify(memberRepository, times(2)).save(any());
     }
 }
